3_net-msg/pyproject.toml
[tool.poetry]
name = "net-msg"
version = "0.1.0"
description = "A simple client-server application."
authors = ["Jean-Philippe Lebœuf <jpl@jean-philippe.leboeuf.name>"]

[tool.poetry.dependencies]
python = "^3.10"
pyzmq = "^23.2.1"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
pylint = "^2.15.0"
pyright = "^1.1.268"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
